+++ src/elements/form-field.html
@@ -7,6 +7,6 @@
   <!-- has a view model this is an custom element with view model -->
   <compose
     if.bind="hasViewModel"
-    model.bind="{attribute: attribute, value: value}"
+    model.bind="{attribute: attribute, value: value, message: message}"
     view-model.bind="view"></compose>
 </template>
