+++ src/elements/form-fields.html
@@ -2,6 +2,7 @@
 
   <form-field
     value.bind="model[attribute.key]"
+    message.bind="messages[attribute.key]"
     attribute.bind="attribute"
     repeat.for="attribute of schema">
   </form-field>
