+++ src/elements/schema-form.html
@@ -1,12 +1,9 @@
 <template>
-
-  <form role="form" submit.delegate="onSubmit()">
-
+  <form role="form">
     <fieldset>
       <content select="schema-form-header"></content>
-      <form-fields model.bind="model" schema.bind="schema"></form-fields>
+      <form-fields messages.bind="messages" model.bind="model" schema.bind="schema"></form-fields>
       <content select="schema-form-footer"></content>
     </fieldset>
-
   </form>
 </template>
