+++ src/elements/schema-form.js
@@ -5,4 +5,7 @@ export class SchemaFormCustomElement {
 
   @bindable({defaultBindingMode: bindingMode.twoWay})
   model
+
+  @bindable
+  messages
 }
